@@ -5,7 +5,7 @@ build-backend = "hatchling.build"
 [project]
 name = "anycorn"
 dynamic = ["version"]
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 description = "A fork of Hypercorn that uses AnyIO"
 authors = [
     { name = "Philip Graham Jones", email = "philip.graham.jones@googlemail.com" },
@@ -19,11 +19,11 @@ classifiers = [
     "Operating System :: OS Independent",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
@@ -33,10 +33,11 @@ readme = "README.md"
 repository = "https://github.com/davidbrochart/anycorn/"
 
 dependencies = [
-    "exceptiongroup >=1.1.0, <2.0; python_version<'3.11'",
+    "exceptiongroup >=1.1.0,<2.0; python_version<'3.11'",
     "h11",
     "h2 >=3.1.0",
-    "anyio >=4.0, <5.0",
+    "hpack",
+    "anyio >=4.0,<5.0",
     "priority",
     "tomli; python_version<'3.11'",
     "typing_extensions; python_version<'3.11'",
@@ -46,7 +47,7 @@ dependencies = [
 
 [project.optional-dependencies]
 h3 = [
-    "aioquic >= 0.9.0, < 1.0",
+    "aioquic >=1.2.0,<2.0.0",
 ]
 test = [
     "httpx",
