@@ -134,7 +134,7 @@ def _build_environ(scope: HTTPScope, body: bytes) -> dict:
         "QUERY_STRING": scope["query_string"].decode("ascii"),
         "SERVER_NAME": server[0],
         "SERVER_PORT": server[1],
-        "SERVER_PROTOCOL": "HTTP/%s" % scope["http_version"],
+        "SERVER_PROTOCOL": f"HTTP/{scope['http_version']}",
         "wsgi.version": (1, 0),
         "wsgi.url_scheme": scope.get("scheme", "http"),
         "wsgi.input": BytesIO(body),
@@ -154,7 +154,7 @@ def _build_environ(scope: HTTPScope, body: bytes) -> dict:
         elif name == "content-type":
             corrected_name = "CONTENT_TYPE"
         else:
-            corrected_name = "HTTP_%s" % name.upper().replace("-", "_")
+            corrected_name = f"HTTP_{name.upper().replace('-', '_')}"
         # HTTPbis say only ASCII chars are allowed in headers, but we latin1 just in case
         value = raw_value.decode("latin1")
         if corrected_name in environ:
