@@ -1,12 +1,14 @@
 from __future__ import annotations
 
-from typing import Awaitable, Callable
+from collections.abc import Awaitable
+from typing import Callable, cast
 
 import h2
 import h2.connection
 import h2.events
 import h2.exceptions
 import priority
+from hpack import HeaderTuple
 
 from ..config import Config
 from ..events import Closed, Event, RawData, Updated
@@ -133,14 +135,14 @@ class H2Protocol:
         self, headers: list[tuple[bytes, bytes]] | None = None, settings: str | None = None
     ) -> None:
         if settings is not None:
-            self.connection.initiate_upgrade_connection(settings)
+            self.connection.initiate_upgrade_connection(settings.encode())
         else:
             self.connection.initiate_connection()
         await self._flush()
         if headers is not None:
             event = h2.events.RequestReceived()
             event.stream_id = 1
-            event.headers = headers
+            event.headers = cast(list[HeaderTuple], headers)
             await self._create_stream(event)
             await self.streams[event.stream_id].handle(EndBody(stream_id=event.stream_id))
         self.task_group.spawn(self.send_task)
@@ -361,7 +363,7 @@ class H2Protocol:
         await self.streams[request.stream_id].handle(
             Request(
                 stream_id=request.stream_id,
-                headers=filter_pseudo_headers(request.headers),
+                headers=filter_pseudo_headers(cast(list[tuple[bytes, bytes]], request.headers)),
                 http_version="2",
                 method=method,
                 raw_path=raw_path,
@@ -392,7 +394,7 @@ class H2Protocol:
         else:
             event = h2.events.RequestReceived()
             event.stream_id = push_stream_id
-            event.headers = request_headers
+            event.headers = cast(list[HeaderTuple], request_headers)
             await self._create_stream(event)
             await self.streams[event.stream_id].handle(EndBody(stream_id=event.stream_id))
             self.keep_alive_requests += 1
