@@ -1,8 +1,10 @@
 from __future__ import annotations
 
-from typing import Any, Callable, Iterable
+from collections.abc import Iterable
+from typing import Any, Callable
 
 import pytest
+
 from anycorn.typing import Scope
 from anycorn.utils import build_and_validate_headers, filter_pseudo_headers, is_asgi, suppress_body
 
