@@ -1,8 +1,9 @@
 from __future__ import annotations
 
+from collections.abc import Awaitable
 from enum import Enum, auto
 from time import time
-from typing import Awaitable, Callable
+from typing import Callable
 from urllib.parse import unquote
 
 from ..config import Config
