@@ -4,6 +4,7 @@ import inspect
 import os
 import socket
 import sys
+from collections.abc import Awaitable, Iterable
 from enum import Enum
 from importlib import import_module
 from multiprocessing.synchronize import Event as EventType
@@ -11,9 +12,7 @@ from pathlib import Path
 from typing import (
     TYPE_CHECKING,
     Any,
-    Awaitable,
     Callable,
-    Iterable,
     Literal,
     cast,
 )
